backend/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.4
psycopg2-binary==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
